split userData into seperate file + adding exerciseTracker
